requirements.txt
# Usage: pip install -r requirements.txt

absl-py==2.1.0
albumentations==1.4.21
asttokens==2.4.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
clearml==1.16.4
contourpy==1.2.1
code-loader==1.0.86
cycler==0.12.1
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
furl==2.1.3
grpcio==1.65.5
idna==3.7
imageio==2.35.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
Markdown==3.7
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
# Base ----------------------------------------
matplotlib>=3.2.2
#numpy>=1.18.5,<1.24.0
#opencv-python>=4.1.1
#
#
#
#
#
#
#
onnx
onnxsim
onnx_graphsurgeon
onnxruntime==1.20.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
#torch>=1.7.0,!=1.12.0
torch==2.4.1
#torchvision>=0.8.1,!=0.13.0
torchvision==0.19.1
tqdm>=4.41.0
protobuf<4.21.3

# Logging -------------------------------------
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1

# wandb

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Export --------------------------------------
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop  # FLOPs computation
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow

openpyxl==3.1.5
orderedmultidict==1.0.1
packaging==24.1
parso==0.8.4
pathlib2==2.3.7.post1
pexpect==4.9.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
scikit-image==0.24.0
scikit-learn==1.5.1
six==1.16.0
slicerator==1.1.0
stack-data==0.6.3
sympy==1.13.2
#tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.8.10
#numpy<1.26.0